- rename sample to golaemSwitch
