+++ vrayGolaemSwitch/vraygolaemSwitch.h
