+++ vrayGolaemSwitch/vraygolaemSwitch.rc
