+++ vrayGolaemSwitch/vraygolaemSwitch_impl.h
